Improve pdsh-sync script
root access is only needed to be able to copy the generated genders file
to the correct location across the cluster.  Let's not run the entire
script as root.  This means that flight asset need only be configured as
the current user, not also as root.

Adjust permissions, ownership of the temporary genders file.  The file
needs to be readable by all users.