flight-pam builder
The standard omnibus builder has not been used. The object file produced
by `flight-pam` needs to integrate with the system pam binary, so
isolating the pam dependency as omnibus does could lead to version
mismatches.  As we're depending on a system pam we may as well depend on
a system libcurl too.